Cleanout game save code
Now follow best practices. Don't store QObject in it, and instead use
roles to expose properties of the game saves. The model is also
imediately updated once a new save is created or deleted.

We need to move more code related to saving from App to there in the
future.

Signed-off-by: Carl Schwan <carl@carlschwan.eu>